requirements.txt
langchain
langchain-groq
langchain-community
langchain-text-splitters
langgraph
qdrant-client
langchain-qdrant
fastapi
uvicorn
python-dotenv
sentence-transformers
pypdf
python-multipart
aiokafka
